Sort input numbers (move sorting)
The number of results is varying based on the sequence of input
numbers used. By always sorting, the way the user chooses the
sequence the number doesn't matter.

The reason for variation in the results seems to be ways of producing '1'
using division.
Example: 50 * 2 / 100 = 1
Based on sorting both results, containing
100 / (50 * 2) and 50 * 2 / 100 are not neccesarily listed, but rather
just one of them.