Implement improved way of running with node.js
Running
  node node-letters.js dunnotwoc

Now, log output (using console.log()) is transferred from the vm
to the user terminal due to that context trick.

The extra debugging code which isn't used in the browser is inserted
here too. It is included due to reading itself into the 'code'.
in_vm is used to determine if the file node-letters.js is running
before the vm is set up or after, inside the vm.