Optimize time consumption.
Now, the new algorithm is faster than the original.

Instead of finding all 9 letter words and then selecting one randomly,
the dictionary now is searched through in a random manner, and the
first found word is used.

It is not known though, if the distribution of selected words is
as random as before.