Rewrite conundrum2.js
A new generator, give_words() is called with two parameters
  - length, the number of letters we want
  - skip, the number of words of that size to be skipped before
    giving the first word.

For instance:
  let words = give_words(9, 0);
  if we are interested in 9 letter words and want the very first word to be generated.

For each new call of words.next(skip).value, a new word is generated.
The skip value is set to zero for not skipping words. If positive,
the skip value is used as the distance to the next word returned.

The generator generates forever, so if the whole dictionary is traversed,
the generator start over.

In generate_conundrum(), the first word produced is one of the 19000 first words,
If no success with that word, just a few more are skipped before testing another
word.

The algorithm should ensure that no 9 letter word is preferred over others
(if possible to solve). However the assumtion that there are 19000 words,
is slightly wrong, and will be fixed.