Change conundrom question slightly
When searching for the four and five letter words to use as
the conundrum question, try make sure that nowhere in the facit
there is a sequence of 4 letters which also appears in the
question. If not possible to achieve, use a weaker question anyway.

Always order the question with the four letter word first.